tool/performance/requirements.txt
duckdb>=1.5.0
opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp>=1.36.0
psycopg[binary]>=3.2.0
